Fix SFTP file upload when targeting an explicit full path
